+++ src/commands/account/create_account/fund_myself_create_account/add_key/autogenerate_new_keypair/mod.rs
@@ -57,17 +57,16 @@ impl From<GenerateKeypairContext> for super::super::AccountPropertiesContext {
 #[strum_discriminants(derive(EnumMessage, EnumIter))]
 /// Save an access key for this account:
 pub enum SaveMode {
-    #[cfg(target_os = "macos")]
     #[strum_discriminants(strum(
-        message = "save-to-macos-keychain   - Save automatically generated key pair to macOS keychain"
+        message = "save-to-keychain   - Save automatically generated key pair to keychain"
     ))]
-    /// Save automatically generated key pair to macOS keychain
-    SaveToMacosKeychain(SignAs),
+    /// Save automatically generated key pair to keychain
+    SaveToKeychain(SignAs),
     #[strum_discriminants(strum(
-        message = "save-to-keychain         - Save automatically generated key pair to the legacy keychain (compatible with JS CLI)"
+        message = "save-to-legacy-keychain         - Save automatically generated key pair to the legacy keychain (compatible with JS CLI)"
     ))]
     /// Save automatically generated key pair to the legacy keychain (compatible with JS CLI)
-    SaveToKeychain(SignAs),
+    SaveToLegacyKeychain(SignAs),
     #[strum_discriminants(strum(
         message = "print-to-terminal        - Print automatically generated key pair in terminal"
     ))]
@@ -93,21 +92,20 @@ impl SaveModeContext {
 
                 move |_signed_transaction, network_config, storage_message| {
                     match scope {
-                        #[cfg(target_os = "macos")]
-                        SaveModeDiscriminants::SaveToMacosKeychain => {
+                        SaveModeDiscriminants::SaveToKeychain => {
                             let key_pair_properties_buf =
                                 serde_json::to_string(&key_pair_properties)?;
-                            *storage_message = crate::common::save_access_key_to_macos_keychain(
+                            *storage_message = crate::common::save_access_key_to_keychain(
                                 network_config.clone(),
                                 &key_pair_properties_buf,
                                 &key_pair_properties.public_key_str,
                                 new_account_id.as_ref(),
                             )?;
                         }
-                        SaveModeDiscriminants::SaveToKeychain => {
+                        SaveModeDiscriminants::SaveToLegacyKeychain => {
                             let key_pair_properties_buf =
                                 serde_json::to_string(&key_pair_properties)?;
-                            *storage_message = crate::common::save_access_key_to_keychain(
+                            *storage_message = crate::common::save_access_key_to_legacy_keychain(
                                 network_config.clone(),
                                 credentials_home_dir.clone(),
                                 &key_pair_properties_buf,
